To Pudim and Back Again, by Bilbo Baggins
Dedicated to Rik, my truest of friends.
May the Ember of Em-Y guide you to glory and fame.

Prologue: An Unexpected Fella

There was once a hobbit named Rik, of the House of Ardo. He was a fairly short
and butch fella (even for a hobbit), with eyes that would change colors at his
will. Eyes that could be both as blue as the sky or green as a marijuana leaf.

But not only of his physical traits was Rik unusual. Since his first breaths
of air, when he was but a little hobbit, fresh out of his mother's womb, Rik
already stirred much trouble due to both his stubborness and unwillingness to
follow rules and norms. Just when the Doctor took baby Rik in his arms, the
little rascal bit the physician's face with such strength that the Doctor,
almost uncounscious with the pain from the bite, dropped the baby to the floor,
legs first, causing his height as it is.

But I digress. For this is not a medical description of Hobbit aberrations,
but rather the story of this young little fella that changed the world as
we know forever.

TO BE CONTINUED...

Next on "To Pudim and Back Again":

Doctor: Oh THAT's a baby with a wrinkled face, ain't you ma lad?!
Rik: Wrinkled is your mom! *bites Doctor's nose and refuses to let go*
Doctor: AHGASKDLJASKLUDIOASJASHDAGHHHHDBASNDJKASND
Rik: That'll show you not to mess with the mother-foca who nailed your mom!

CUTS TO BLACK -- DRAMATIC MUSIC STARTS PLAYING -- CREDITS ROLL